requirements.txt
discord.py==2.4.0
PyNaCl==1.5.0
yt-dlp==2026.2.21
